+++ src/components/App.jsx
@@ -22,7 +22,7 @@ import { setSession, clearSession } from '../session'
 import Onboarding from './Onboarding'
 import Cheatsheet from './Cheatsheet'
 import { routes } from '../constants/routes'
-import { isFeatureEnabled } from '../constants/featureFlags'
+import packageInfo from '../../package.json'
 
 export default function App() {
   const { t } = useTranslation()
@@ -37,8 +37,7 @@ export default function App() {
   const [showAlphaWarning, setShowAlphaWarning] = useState(false)
   const [showCheatsheet, setShowCheatsheet] = useState(false)
 
-  const cheatsheetEnabled = currentWallet && isFeatureEnabled('cheatsheet')
-  const scheduleEnabled = currentWallet && isFeatureEnabled('schedule')
+  const cheatsheetEnabled = currentWallet
 
   const startWallet = useCallback(
     (name, token) => {
@@ -125,7 +124,7 @@ export default function App() {
                 <Route path={routes.home} element={<Wallets startWallet={startWallet} stopWallet={stopWallet} />} />
                 {currentWallet && (
                   <>
-                    {scheduleEnabled && <Route path={routes.jam} element={<Jam />} />}
+                    <Route path={routes.jam} element={<Jam />} />
                     <Route path={routes.send} element={<Send />} />
                     <Route path={routes.earn} element={<Earn />} />
                     <Route path={routes.receive} element={<Receive />} />
@@ -185,6 +184,17 @@ export default function App() {
             )}
           </div>
           <div className="d-flex flex-1 order-2 justify-content-end align-items-center gap-1">
+            <div className="warning-hint text-start text-secondary d-none d-md-block pe-1">
+              <a
+                href="https://github.com/joinmarket-webui/joinmarket-webui/tags"
+                target="_blank"
+                rel="noopener noreferrer"
+                className="d-flex align-items-center text-secondary"
+              >
+                {' '}
+                v{packageInfo.version}
+              </a>
+            </div>
             <div className="d-flex gap-2 pe-2">
               <a
                 href="https://github.com/joinmarket-webui/joinmarket-webui"